added more functionality to guestdetails modal
